requirements.txt
cffi==1.11.5
Click==7.0
Flask==1.0.2
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
numpy==1.16.0
pkg-resources==0.0.0
pycparser==2.19
PySoundFile==0.9.0.post1
sounddevice==0.3.12
Werkzeug==0.14.1
